requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
pytest-asyncio==0.20.3
pytest==7.2.1
discord~=2.1.0
aiohttp~=3.8.3